fixing contact page and the layout page
